+++ packages/backend/test-fns/jest.config.js
@@ -1,10 +1,9 @@
 // test-fns/jest.config.js
 
-import defaults from '../jest.config.default.js'
+import opts from '../jest.config.default.js'
 
-const opts = { ...defaults,
+export default { ...opts,
   // Jest default is 5000. Here to help debugging, if some 'fns' test times out.
   testTimeout: 3000,
+  globalSetup: "./setup.jest.js"
 };
-
-export default opts;