+++ packages/backend/test-rules/invitesC.test.js
@@ -5,7 +5,9 @@ import { strict as assert } from 'assert'
 
 import { test, expect, describe, beforeAll } from '@jest/globals'
 
-import { dbAuth, FieldValue } from 'firebase-jest-testing/firestoreReadOnly'
+import { collection, serverTimestamp } from 'firebase-jest-testing/firestoreRules'
+
+const SERVER_TIMESTAMP = serverTimestamp();
 
 const anyDate = new Date();   // a non-server date
 
@@ -13,45 +15,34 @@ describe("'/invites' rules", () => {
   let unauth_invitesC, auth_invitesC, abc_invitesC, def_invitesC;
 
   beforeAll( () => {
-    try {
-      const coll = dbAuth.collection('invites');   // root collection
+    const coll = collection('invites');   // root collection
 
-      unauth_invitesC = coll.as(null);
-      auth_invitesC = coll.as({uid:'_'});
-      abc_invitesC = coll.as({uid:'abc'});
-      def_invitesC = coll.as({uid:'def'});
-
-      assert(unauth_invitesC && auth_invitesC && abc_invitesC && def_invitesC);
-    }
-    catch (err) {
-      // tbd. How to cancel the tests if we end up here? #help
-      console.error( "Failed to initialize the Firebase database: ", err );
-      throw err;
-    }
+    unauth_invitesC = coll.as(null);
+    auth_invitesC = coll.as({uid:'_'});
+    abc_invitesC = coll.as({uid:'abc'});
+    def_invitesC = coll.as({uid:'def'});
   });
 
   //--- InvitesC read rules ---
 
-  test('no-one should be able to read', async () => {
+  test('no-one should be able to read', () => Promise.all([
 
-    await Promise.all([
       expect( unauth_invitesC.get() ).toDeny(),   // unauthenticated
       expect( auth_invitesC.get() ).toDeny(),     // valid user (trying to list the invites)
 
       expect( abc_invitesC.get("a@b.com:1") ).toDeny()   // the one who's created an invite
-    ]);
-  });
+  ]));
 
   //--- InvitesC create rules ---
 
-  test('only a member of a project can invite; only author can invite as-author', async () => {
+  test('only a member of a project can invite; only author can invite as-author', () => {
     const template = { email: "aa@b.com", project: "1" };
     const dGen = (uid, asAuthor) => ({ ...template,
-      asAuthor: asAuthor, by: uid, at: FieldValue.serverTimestamp()
+      asAuthor: asAuthor, by: uid, at: SERVER_TIMESTAMP
     });
     const id = `${template.email}:${template.project}`;
 
-    await Promise.all([
+    return Promise.all([
       expect( abc_invitesC.doc(id).set( dGen("abc",true )) ).toAllow(),   // author can invite as-author
       expect( abc_invitesC.doc(id).set( dGen("abc",false )) ).toAllow(),  // ..or as collaborator
 
@@ -79,13 +70,13 @@ describe("'/invites' rules", () => {
   });
   ***/
 
-  test('validity: server time; identifying oneself; \'email:project\' as id', async () => {
+  test('validity: server time; identifying oneself; \'email:project\' as id', () => {
     const d = {
       email: "aaa@b.com",
       project: "1",
       asAuthor: false,
       by: "abc",
-      at: FieldValue.serverTimestamp()
+      at: SERVER_TIMESTAMP
     };
     const id = `${d.email}:${d.project}`;
     const badId = "blah";
@@ -93,7 +84,7 @@ describe("'/invites' rules", () => {
     const dBadTime = { ...d, at: anyDate };
     const dNotMe = { ...d, by: "zwho" };
 
-    await Promise.all([
+    return Promise.all([
       expect( abc_invitesC.doc(id).set(d) ).toAllow(),   // re-check valid data gets through
       expect( abc_invitesC.doc(id).set(dBadTime) ).toDeny(),
       expect( abc_invitesC.doc(id).set(dNotMe) ).toDeny(),