+++ packages/backend/test-rules/jest.config.js
@@ -1,10 +1,7 @@
 // test-rules/jest.config.js
 
-import defaults from '../jest.config.default.js'
+import opts from '../jest.config.default.js'
 
-const opts = { ...defaults,
-  // Load docs, once at the beginning of the tests.
-  globalSetup: "./setup.jest.cjs"
+export default { ...opts,
+  globalSetup: "./setup.jest.js"
 };
-
-export default opts;